refactored `ExposedLinuxFile` into `.test` namespace.
